zod validation on server and client side

+++ src/actions/actions.ts
@@ -3,16 +3,23 @@
 import prisma from "@/lib/db";
 import { TPetEssential } from "@/lib/types";
 import { sleep } from "@/lib/utils";
+import { petFormZodValidation, petIdValidation } from "@/lib/validations";
 import { Pet } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(newPet: TPetEssential) {
+export async function addPet(newPet: unknown) {
+  //ставим unknown так как у unknow нет редактируеміх свойст .нельзя записать pet.age=rm-rf .
   await sleep();
+  // проверяем точно ли соответствует наша схема тому что нужно
+  const validatedPet = petFormZodValidation.safeParse(newPet);
+  if (!validatedPet.success) {
+    return { message: "Something Wrong" };
+  }
   //в функцию приходит formData из actions формі .Достаем ее
   // По факту не нужен отдельній ендпоинт .Тут он и есть для связи с ORM
   try {
     await prisma.pet.create({
-      data: newPet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return { message: "Something Wrong" };
@@ -26,14 +33,19 @@ export async function getPets() {
   return await prisma.pet.findMany();
 }
 
-export async function editPet(petID: Pet["id"], newPet: TPetEssential) {
+export async function editPet(petID: unknown, newPet: unknown) {
   await sleep();
   //в функцию приходит formData из actions формі .Достаем ее
   // По факту не нужен отдельній ендпоинт .Тут он и есть для связи с ORM
+  const validateidPet = petIdValidation.safeParse(petID);
+  const validatedPet = petFormZodValidation.safeParse(newPet);
+  if (!validatedPet.success || !validateidPet.success) {
+    return { message: "Something Wrong" };
+  }
   try {
     await prisma.pet.update({
-      where: { id: petID },
-      data: newPet,
+      where: { id: validateidPet.data },
+      data: validatedPet.data,
     });
   } catch (error) {
     return { message: "Something Wrong" };
@@ -43,11 +55,16 @@ export async function editPet(petID: Pet["id"], newPet: TPetEssential) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petID: Pet["id"]) {
+export async function deletePet(petID: unknown) {
   await sleep();
+
+  const validateidPet = petIdValidation.safeParse(petID);
+  if (!validateidPet.success) {
+    return { message: "Something Wrong" };
+  }
   try {
     await prisma.pet.delete({
-      where: { id: petID },
+      where: { id: validateidPet.data },
     });
   } catch (error) {
     return { message: "Something Wrong" };
